+++ packages/@dicebear/adventurer-neutral/src/colors/index.ts
@@ -1,0 +1,1 @@
+export { skin } from './skin';