+++ packages/@dicebear/adventurer-neutral/src/components/index.ts
@@ -1,0 +1,4 @@
+export { eyes } from './eyes';
+export { eyebrows } from './eyebrows';
+export { mouth } from './mouth';
+export { accessoires } from './accessoires';