+++ packages/@dicebear/adventurer/src/colors/index.ts
@@ -1,0 +1,2 @@
+export { skin } from './skin';
+export { hair } from './hair';