+++ packages/@dicebear/adventurer/src/components/index.ts
@@ -1,0 +1,5 @@
+export { eyes } from './eyes';
+export { eyebrows } from './eyebrows';
+export { mouth } from './mouth';
+export { accessoires } from './accessoires';
+export { hair } from './hair';